Move verifyConditions to utils, no functional change
